+++ CHANGELOG.md
@@ -4,9 +4,11 @@
 
 ### Added
 - Проверка рефереров по спискам (Allow/Block/Captcha)
+- ASNChecker: добавлен список captcha_asn
 
 ### Changed
 - Удалены настройки direct = , referer =. В место них проверки по спискам
+- ASNChecker: секции [asn_checker] и [asn_whilelist] объединены в одну [asn_checker]
 
 ### Fixed
 -